+++ htag/runners/devapp.py
@@ -48,7 +48,7 @@ class DevApp(Starlette):
 
             async def on_receive(this, websocket, data):
                 className = data["class"]
-                actions = await self.hrenderers[className].interact(data["id"],data["method"],data["args"],data["kargs"])
+                actions = await self.hrenderers[className].interact(data["id"],data["method"],data["args"],data["kargs"],data["event"])
                 await websocket.send_text( json.dumps(actions) )
 
         Starlette.__init__(self,debug=True, routes=[